raftstore: extract Callback to tikv_misc
