+++ components/tikv_misc/src/raftstore_callback.rs
@@ -1,0 +1,68 @@
+use std::boxed::FnBox;
+use std::fmt;
+use kvproto::raft_cmdpb::RaftCmdResponse;
+use super::region_snapshot::RegionSnapshot;
+
+#[derive(Debug, Clone)]
+pub struct ReadResponse {
+    pub response: RaftCmdResponse,
+    pub snapshot: Option<RegionSnapshot>,
+}
+
+#[derive(Debug)]
+pub struct WriteResponse {
+    pub response: RaftCmdResponse,
+}
+
+pub type ReadCallback = Box<dyn FnBox(ReadResponse) + Send>;
+pub type WriteCallback = Box<dyn FnBox(WriteResponse) + Send>;
+
+/// Variants of callbacks for `Msg`.
+///  - `Read`: a callbak for read only requests including `StatusRequest`,
+///         `GetRequest` and `SnapRequest`
+///  - `Write`: a callback for write only requests including `AdminRequest`
+///          `PutRequest`, `DeleteRequest` and `DeleteRangeRequest`.
+pub enum Callback {
+    /// No callback.
+    None,
+    /// Read callback.
+    Read(ReadCallback),
+    /// Write callback.
+    Write(WriteCallback),
+}
+
+impl Callback {
+    pub fn invoke_with_response(self, resp: RaftCmdResponse) {
+        match self {
+            Callback::None => (),
+            Callback::Read(read) => {
+                let resp = ReadResponse {
+                    response: resp,
+                    snapshot: None,
+                };
+                read(resp);
+            }
+            Callback::Write(write) => {
+                let resp = WriteResponse { response: resp };
+                write(resp);
+            }
+        }
+    }
+
+    pub fn invoke_read(self, args: ReadResponse) {
+        match self {
+            Callback::Read(read) => read(args),
+            other => panic!("expect Callback::Read(..), got {:?}", other),
+        }
+    }
+}
+
+impl fmt::Debug for Callback {
+    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match *self {
+            Callback::None => write!(fmt, "Callback::None"),
+            Callback::Read(_) => write!(fmt, "Callback::Read(..)"),
+            Callback::Write(_) => write!(fmt, "Callback::Write(..)"),
+        }
+    }
+}
